Move transfer mod into precompiles mod
