+++ crates/celo-revm/src/lib.rs
@@ -8,7 +8,6 @@ extern crate alloc as std;
 pub mod chain_info;
 pub mod evm;
 pub mod precompiles;
-pub mod transfer;
 pub mod tx;
 
 pub use evm::CeloEvm;
